requirements.txt
adrf==0.1.6
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
anyio==4.4.0
asgiref==3.7.2
async-property==0.2.2
attrs==23.2.0
boto3
botocore
cement==2.10.14
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
current==0.4.1
Django==5.0.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-filter==23.5
django-quill-editor==0.1.40
django-storages==1.14.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dotsi==0.0.3
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
jmespath==1.0.1
Markdown==3.5.2
multidict==6.0.5
packaging==24.0
pathspec==0.10.1
phonenumbers==8.13.46
pillow
# psycopg2==2.9.9
# psycopg2-binary==2.9.9
PyJWT==2.8.0
pypiwin32==223; platform_system == "Windows"
python-dateutil==2.9.0.post0
pytz==2024.1
pywin32==308; platform_system == "Windows"
PyYAML==6.0.2
requests==2.31.0
s3transfer==0.10.1
semantic-version==2.10.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
termcolor==2.5.0
twilio==9.0.0
tzdata==2024.1
urllib3==1.26.20
uvicorn==0.30.1
wcwidth==0.2.13
yarl==1.9.4
django-cors-headers